+++ tests/Generator/MakeModelTest.php
@@ -106,9 +106,11 @@ it('normalizes generated model to pascal case', function ($given, $normalized) {
     expect(file_exists($expectedModelPath))->toBeTrue();
 })->with('makeModelInputs');
 
-it('generates the base model when possible', function () {
+it('generates the base model when possible', function ($baseModelClass, $baseModelPath) {
     $modelName = Str::studly(fake()->word());
     $domain = Str::studly(fake()->word());
+
+    Config::set('ddd.base_model', $baseModelClass);
 
     $expectedModelPath = base_path(implode('/', [
         config('ddd.paths.domains'),
@@ -123,23 +125,36 @@ it('generates the base model when possible', function () {
 
     expect(file_exists($expectedModelPath))->toBeFalse();
 
-    // This currently only tests for the default base model
-    $expectedBaseModelPath = base_path(config('ddd.paths.domains').'/Shared/Models/BaseModel.php');
+    $expectedBaseModelPath = base_path($baseModelPath);
 
     if (file_exists($expectedBaseModelPath)) {
         unlink($expectedBaseModelPath);
     }
 
-    // Todo: should bypass base model creation if
-    // a custom base model is being used.
-    // $baseModel = config('ddd.base_model');
-
-    expect(file_exists($expectedBaseModelPath))->toBeFalse();
+    expect(file_exists($expectedBaseModelPath))->toBeFalse("{$baseModelPath} expected not to exist.");
 
     Artisan::call("ddd:model {$domain} {$modelName}");
 
-    expect(file_exists($expectedBaseModelPath))->toBeTrue();
-});
+    expect(file_exists($expectedBaseModelPath))->toBeTrue("{$baseModelPath} expected to exist.");
+})->with([
+    ['Domain\\Shared\\Models\\BaseModel', 'src/Domain/Shared/Models/BaseModel.php'],
+    ['Domain\\Infrastructure\\Models\\BaseModel', 'src/Domain/Infrastructure/Models/BaseModel.php'],
+]);
+
+it('will not create a base model if the configured base model is out of scope', function ($baseModel) {
+    Config::set('ddd.base_model', $baseModel);
+
+    expect(class_exists($baseModel))->toBeFalse();
+
+    Artisan::call('ddd:model Fruits Lemon');
+
+    expect(Artisan::output())
+        ->toContain("Configured base model {$baseModel} doesn't exist.")
+        ->not->toContain("Generating {$baseModel}");
+})->with([
+    ['Illuminate\\Database\\Eloquent\\NonExistentModel'],
+    ['OtherVendor\\OtherPackage\\Models\\OtherModel'],
+]);
 
 it('skips base model creation if configured base model already exists', function ($baseModel) {
     Config::set('ddd.base_model', $baseModel);
@@ -148,10 +163,12 @@ it('skips base model creation if configured base model already exists', function
 
     Artisan::call('ddd:model Fruits Lemon');
 
-    expect(Artisan::output())->not->toContain("Base model {$baseModel} doesn't exist, generating...");
+    expect(Artisan::output())
+        ->not->toContain("Configured base model {$baseModel} doesn't exist.")
+        ->not->toContain("Generating {$baseModel}");
 })->with([
-    ['Illuminate\Database\Eloquent\Model'],
-    ['Lunarstorm\LaravelDDD\Models\DomainModel'],
+    ['Illuminate\\Database\\Eloquent\\Model'],
+    ['Lunarstorm\\LaravelDDD\\Models\\DomainModel'],
 ]);
 
 it('shows meaningful hints when prompting for missing input', function () {